Added changes to profile page
